module_migration_dm/requirements.txt
sqlalchemy==1.4.54
alembic==1.16.2
psycopg2-binary==2.9.10
pytest==8.4.1
python-dotenv==1.1.1
